pyproject.toml
[tool.poetry]
name = "bananagan"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "bananagan", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.3.1"
torchvision = "^0.18.1"
matplotlib = "^3.9.1"
numpy = "^2.0.1"
opencv-python = "^4.10.0.84"
pillow = "^10.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
